services/evaluator/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
PyYAML==6.0.1
pgvector

scikit-learn==1.3.2
